move setting view models in own dir / file
