+++ src/domain/session/settings/SessionBackupViewModel.js
@@ -1,5 +1,4 @@
 /*
-Copyright 2020 Bruno Windels <bruno@windels.cloud>
 Copyright 2020 The Matrix.org Foundation C.I.C.
 
 Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,63 +14,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {ViewModel} from "../ViewModel.js";
+import {ViewModel} from "../../ViewModel.js";
 
-export class SettingsViewModel extends ViewModel {
-    constructor(options) {
-        super(options);
-        this._updateService = options.updateService;
-        const session = options.session;
-        this._session = session;
-        this._sessionBackupViewModel = this.track(new SessionBackupViewModel(this.childOptions({session})));
-        this._closeUrl = this.urlCreator.urlUntilSegment("session");
-    }
-
-    get closeUrl() {
-        return this._closeUrl;
-    }
-
-    get fingerprintKey() {
-        const key = this._session.fingerprintKey;
-        const partLength = 4;
-        const partCount = Math.ceil(key.length / partLength);
-        let formattedKey = "";
-        for (let i = 0; i < partCount; i += 1) {
-            formattedKey += (formattedKey.length ? " " : "") + key.slice(i * partLength, (i + 1) * partLength);
-        }
-        return formattedKey;
-    }
-
-    get deviceId() {
-        return this._session.deviceId;
-    }
-
-    get userId() {
-        return this._session.userId;
-    }
-
-    get version() {
-        if (this._updateService) {
-            return `${this._updateService.version} (${this._updateService.buildHash})`;
-        }
-        return this.i18n`development version`;
-    }
-
-    checkForUpdate() {
-        this._updateService?.checkForUpdate();
-    }
-
-    get showUpdateButton() {
-        return !!this._updateService;
-    }
-
-    get sessionBackupViewModel() {
-        return this._sessionBackupViewModel;
-    }
-}
-
-
-class SessionBackupViewModel extends ViewModel {
+export class SessionBackupViewModel extends ViewModel {
     constructor(options) {
         super(options);
         this._session = options.session;
